Bump CI Rust toolchain to 1.91.0 and fix clippy warnings
soroban-sdk 27.0.2 requires rustc 1.91.0+ (this workspace's Cargo.toml
already resolves to that version once the stellar-contracts dependency
pin was fixed on this branch). CI's rust-toolchain.toml and the
per-workflow toolchain override were still pinned to 1.89.0, so
Contracts (Rust) and Format/Lint/Test/Build Soroban Contracts failed
before a single line of contract code even ran.

Also:
- Building for wasm32v1-none now requires either `stellar contract
  build` (stellar-cli v25.2.0+) or the SOROBAN_SDK_BUILD_SYSTEM_SUPPORTS_SPEC_SHAKING_V2=1
  env var that CLI sets internally — added it to every workflow's wasm
  build step so plain `cargo build` keeps working in CI.
- deploy.yml's contract-build job targeted the deprecated
  wasm32-unknown-unknown, which soroban-sdk 27 explicitly rejects on
  rustc 1.82+; switched it to wasm32v1-none to match the other workflows.
- Fixed the 7 clippy warnings that -D warnings promotes to errors under
  1.91.0's newer lint set (needless double-references, unnecessary
  &-borrows on Copy types, a doc-comment/item gap, a redundant cast, and
  two assert!-on-const-expression lints in test-only code).

Verified locally with the same toolchain/flags CI uses: fmt --check,
clippy --all-targets -D warnings, test --workspace, and
cargo build --release --target wasm32v1-none all pass.

+++ contracts/prompt-hash/src/ttl_test.rs
@@ -53,7 +53,7 @@ mod tests {
         let last_extended = 100_000u64;
 
         let remaining = get_time_remaining(current, last_extended, max_ttl);
-        let expected = (last_extended as u64 + max_ttl as u64) - current;
+        let expected = (last_extended + max_ttl as u64) - current;
 
         assert_eq!(remaining, expected);
     }
@@ -78,6 +78,7 @@ mod tests {
     }
 
     #[test]
+    #[allow(clippy::assertions_on_constants)]
     fn test_batch_size_respected() {
         assert!(MAX_RENEWAL_BATCH_SIZE > 0, "Batch size must be positive");
         assert!(
